refactor: relocate models to app\/Models
