+++ app/Models/Assessment.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Assessment extends Model
+{
+    //
+}
