+++ app/Models/Course.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Models\Models;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class SlotBatch extends Model
+class Course extends Model
 {
     //
 }
