--- app/Models/Models/RegulatoryAction.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Models\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class RegulatoryAction extends Model
-{
-    //
-}
