+++ app/Models/SlotBatch.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Models\Models;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Student extends Model
+class SlotBatch extends Model
 {
     //
 }
