+++ app/Models/Student.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Models\Models;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Evidence extends Model
+class Student extends Model
 {
     //
 }
